test(ocr): make the third routing copy visible to the parity test (RED)
The OCR model routing had three live copies, and tests/unit/translate-core-parity
only watched two of them: the TS one (API routes, Lambda) and translate-core.mjs
(translation workers). The third lived as a private LATIN_SCRIPT_LANGS_FOR_LITE
inside scripts/workers/pipeline-orchestrator.mjs, which no test can import —
that file calls run() at module load, so importing it would start the worker.
A fourth, undocumented copy sat in scripts/migration/backfill-ocr-near-complete.mjs.

This commit only MOVES the orchestrator's copy, verbatim (malay/ms/msa still in
the list), into scripts/lib/ocr-routing.mjs so it becomes importable, points
both scripts at it, and extends the parity suite to run all three live
implementations over the same probes.

The suite is red at this commit: 4 failures, all malay/ms/msa routing to
flash-lite for OCR while routing to flash for translation. That is the drift,
and it is the negative control for the fix in the next commit — the guard is
worthless unless it can be seen to fail.

Refs: handoff 2026-09-04 (ocr-routing-drift), #3725
